Refactor and enhance export functionality - v0.5.6
- Modularized create_kml_process for improved maintainability.
- Refined filename input logic to fix the issue of asking for the filename twice.
- Ensured export functions are solely responsible for exporting content.
- Implemented ID assignment for polygon points and updated construction sequence creation.
- Removed invalid monument data from JSON when monument is not defined.
- Removed redundant code for file existence checks and path generation.
- Updated the handling of points in export_kml to ensure correct polygon closure.
- Removed monument placemark when monument data is not valid.
- Renamed 'initial' to 'tie_point' for better clarity in JSON output.
- Reorganized code for a clearer separation of concerns across modules.
