+++ VERSION
@@ -1,1 +1,1 @@
-0.5.5
+0.5.6
